+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   - Improve internal store collection performances (x15).
   - Lunaris do not try to compute anymore POST->DELETE and UPDATE->DELETE offline transactions.
   - Add lazy loading for stores. Use `isLazyLoad = true` in stores to activate it.
+  - Change `lunaris.getOne` to use primary key: `lunaris.getOne(store, value, isPrimaryKey)`.
 - Fix:
   - Improve offline app.
   - In offline or offline mode, generated urls were not encoded.